pyproject.toml
[build-system]
requires = ["setuptools>=61.2", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dosage"
description = "a comic strip downloader and archiver"
readme = "README.md"
maintainers = [{name = "Tobias Gruetzmacher", email = "tobias-dosage@23.gs"}]
license = {text = "MIT License"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Multimedia :: Graphics",
]
keywords = ["comic", "webcomic", "downloader", "archiver", "crawler"]
requires-python = ">=3.7"
dependencies = [
    "colorama",
    "imagesize",
    "lxml>=4.0.0",
    "platformdirs",
    "requests>=2.0",
    "cached_property;python_version<'3.8'",
    "importlib_metadata;python_version<'3.8'",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://dosage.rocks"
Code = "https://github.com/webcomics/dosage"
"Issue tracker" = "https://github.com/webcomics/dosage/issues"

[project.optional-dependencies]
bash = ["argcomplete"]
css = ["cssselect"]
dev = [
    "allure-pytest",
    "pytest-cov",
    "pytest-xdist",
    "responses",
    "setup-cfg-fmt",
]
lint = [
    "flake8<6",
    "flake8-2020",
    "flake8-breakpoint",
    "flake8-bugbear",
    "flake8-coding",
    "flake8-commas",
    "flake8-comprehensions",
    "flake8-eradicate",
    "flake8-fixme",
    "flake8-functions",
    "flake8-future-import",
    "flake8-logging-format",
    "flake8-noqa",
    "Flake8-pyproject",
    "flake8-pytest",
    "flake8-pytest-style",
]

[project.scripts]
dosage = "dosagelib.cmd:main"

[project.entry-points.pyinstaller40]
hook-dirs = "dosagelib.__pyinstaller:get_hook_dirs"

[tool.setuptools]
platforms = ["Any"]
license-files = ["COPYING"]

[tool.setuptools.packages]
find = {namespaces = false}

[tool.distutils.bdist_wheel]
universal = true

[tool.setuptools_scm]

[tool.flake8]
max_line_length = 100
ignore = [
    'E127', 'E128',  # hanging instead of visual indent
    'FI18', 'FI58',  # ignore __future__ "annotations" for now
    'W504'  # line breaks after binary operator
]
noqa-require-code = true
no-accept-encodings = true
min-version = "3.7"
extend-exclude = [
    '.venv',
    'build',
]
